refactor(getProfile): created profileUtils for better logics modularization
